requirements.txt

fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
openai-agents
requests
pypdf
openpyxl
pandas
numpy
cuid

eval-type-backport

psycopg2-binary>=2.9.9

e2b_code_interpreter[sandbox]